pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "research-agent"
version = "0.1.0"
description = "LangGraph deep research agent with terminal interaction and ReAct search"
authors = ["Research Agent <research@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.55"
langchain = "^0.3.10"
langchain-openai = "^0.2.11"
langchain-anthropic = "^0.2.7"
tavily-python = "^0.5.0"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
ruff = "^0.0.280"